--- Backend/.env
@@ -1,1 +1,0 @@
-GEMINI_API_KEY = "AIzaSyANkDlAna4DgcQNu-nIka8EiJrKOW1hlzE"
